DQ Memory Quota 2.0: move the operator memory quota interface into its own library
ydb/library/yql/dq/actors/compute peered ydb/library/yql/dq/comp_nodes only to
reach the ~30-line TLS holder dq_operator_memory_quota.{h,cpp}. That target is a
shim that unconditionally peers comp_nodes/llvm16, so every consumer of the
compute actor library pulled in the LLVM comp-node variant, and any target that
wants no_llvm plus the compute actor would fail to configure (both variants
declare PROVIDES(YDB_DQ_COMP_NODES)).

Move the two files into a dependency-free LIBRARY at
comp_nodes/operator_memory_quota and peer that instead. The .cpp is dropped from
ya.make.inc ORIG_SOURCES (it was copied into both llvm16 and no_llvm) and the new
library is peered from there, so the TLS variable is defined exactly once.

actors/compute/ut gains an explicit comp_nodes peerdir: the async/sync compute
actor tests include yql_common_dq_factory.h and relied on the removed transitive
dependency.
