+++ ydb/core/kqp/tools/combiner_perf/dq_combine_vs.cpp
@@ -7,7 +7,7 @@
 #include "kqp_setup.h"
 
 #include <ydb/library/yql/dq/comp_nodes/ut/utils/preallocated_spiller.h>
-#include <ydb/library/yql/dq/comp_nodes/dq_operator_memory_quota.h>
+#include <ydb/library/yql/dq/comp_nodes/operator_memory_quota/dq_operator_memory_quota.h>
 #include <ydb/library/yql/dq/comp_nodes/dq_rh_hash.h>
 #include <yql/essentials/minikql/comp_nodes/ut/mkql_computation_node_ut.h>
 #include <yql/essentials/minikql/computation/mkql_computation_node_holders.h>
