+++ ydb/library/yql/dq/actors/compute/ut/dq_compute_memory_quota_ut.cpp
@@ -1,6 +1,6 @@
 #include <ydb/library/yql/dq/actors/compute/dq_compute_actor.h>
 #include <ydb/library/yql/dq/actors/compute/dq_compute_memory_quota.h>
-#include <ydb/library/yql/dq/comp_nodes/dq_operator_memory_quota.h>
+#include <ydb/library/yql/dq/comp_nodes/operator_memory_quota/dq_operator_memory_quota.h>
 
 #include <yql/essentials/minikql/mkql_alloc.h>
