+++ ydb/library/yql/dq/actors/compute/ut/ya.make
@@ -28,6 +28,7 @@ PEERDIR(
     ydb/library/yql/dq/actors/compute/ut/proto
     ydb/library/yql/dq/actors/input_transforms
     ydb/library/yql/dq/actors/task_runner
+    ydb/library/yql/dq/comp_nodes
     ydb/library/yql/dq/runtime/streaming
     ydb/library/yql/dq/tasks
     ydb/library/yql/dq/transform
