+++ ydb/library/yql/dq/actors/compute/ya.make
@@ -24,7 +24,7 @@ PEERDIR(
     ydb/library/yql/dq/actors/compute/events
     ydb/library/yql/dq/actors/spilling
     ydb/library/yql/dq/common
-    ydb/library/yql/dq/comp_nodes
+    ydb/library/yql/dq/comp_nodes/operator_memory_quota
     ydb/library/yql/dq/proto
     ydb/library/yql/dq/runtime
     ydb/library/yql/dq/runtime/streaming
