+++ ydb/library/yql/dq/comp_nodes/operator_memory_quota/ya.make
@@ -1,0 +1,7 @@
+LIBRARY()
+
+SRCS(
+    dq_operator_memory_quota.cpp
+)
+
+END()
