+++ ydb/library/yql/dq/comp_nodes/ut/dq_operator_memory_quota_ut.cpp
@@ -1,4 +1,4 @@
-#include <ydb/library/yql/dq/comp_nodes/dq_operator_memory_quota.h>
+#include <ydb/library/yql/dq/comp_nodes/operator_memory_quota/dq_operator_memory_quota.h>
 
 #include <library/cpp/testing/unittest/registar.h>
